+++ conans/client/graph/proxy.py
@@ -35,26 +35,26 @@ class ConanProxy(object):
 
         # NOT in disk, must be retrieved from remotes
         if not os.path.exists(conanfile_path):
-            ref_remote = self._download_recipe(reference, output, remote_name)
+            remote, new_ref = self._download_recipe(reference, output, remote_name)
             status = RECIPE_DOWNLOADED
-            return conanfile_path, status, ref_remote
+            return conanfile_path, status, remote, new_ref
 
-        ref_remote = self._registry.get_ref(reference)
+        remote = self._registry.get_recipe_remote(reference)
         check_updates = check_updates or update
         # Recipe exists in disk, but no need to check updates
         if not check_updates:
             status = RECIPE_INCACHE
             log_recipe_got_from_local_cache(reference)
             self._recorder.recipe_fetched_from_cache(reference)
-            return conanfile_path, status, ref_remote
+            return conanfile_path, status, remote, reference
 
         named_remote = self._registry.remote(remote_name) if remote_name else None
-        update_remote = named_remote or ref_remote
+        update_remote = named_remote or remote
         if not update_remote:
             status = RECIPE_NO_REMOTE
             log_recipe_got_from_local_cache(reference)
             self._recorder.recipe_fetched_from_cache(reference)
-            return conanfile_path, status, None
+            return conanfile_path, status, None, reference
 
         try:  # get_conan_manifest can fail, not in server
             upstream_manifest = self._remote_manager.get_conan_manifest(reference, update_remote)
@@ -62,7 +62,7 @@ class ConanProxy(object):
             status = RECIPE_NOT_IN_REMOTE
             log_recipe_got_from_local_cache(reference)
             self._recorder.recipe_fetched_from_cache(reference)
-            return conanfile_path, status, update_remote
+            return conanfile_path, status, update_remote, reference
 
         export = self._client_cache.export(reference)
         read_manifest = FileTreeManifest.load(export)
@@ -71,7 +71,7 @@ class ConanProxy(object):
                 if update:
                     DiskRemover(self._client_cache).remove_recipe(reference)
                     output.info("Retrieving from remote '%s'..." % update_remote.name)
-                    self._remote_manager.get_recipe(reference, update_remote)
+                    reference = self._remote_manager.get_recipe(reference, update_remote)
                     self._registry.set_ref(reference, update_remote)
                     status = RECIPE_UPDATED
                 else:
@@ -83,35 +83,36 @@ class ConanProxy(object):
 
         log_recipe_got_from_local_cache(reference)
         self._recorder.recipe_fetched_from_cache(reference)
-        return conanfile_path, status, update_remote
+        return conanfile_path, status, update_remote, reference
 
     def _download_recipe(self, conan_reference, output, remote_name):
         def _retrieve_from_remote(the_remote):
             output.info("Trying with '%s'..." % the_remote.name)
-            self._remote_manager.get_recipe(conan_reference, the_remote)
-            self._registry.set_ref(conan_reference, the_remote)
+            new_reference = self._remote_manager.get_recipe(conan_reference, the_remote)
+            self._registry.set_ref(new_reference, the_remote)
             self._recorder.recipe_downloaded(conan_reference, the_remote.url)
+            return new_reference
 
         if remote_name:
             output.info("Not found, retrieving from server '%s' " % remote_name)
-            ref_remote = self._registry.remote(remote_name)
+            remote = self._registry.remote(remote_name)
         else:
-            ref_remote = self._registry.get_ref(conan_reference)
-            if ref_remote:
-                output.info("Retrieving from predefined remote '%s'" % ref_remote.name)
+            remote = self._registry.get_recipe_remote(conan_reference)
+            if remote:
+                output.info("Retrieving from predefined remote '%s'" % remote.name)
 
-        if ref_remote:
+        if remote:
             try:
-                _retrieve_from_remote(ref_remote)
-                return ref_remote
+                new_ref = _retrieve_from_remote(remote)
+                return remote, new_ref
             except NotFoundException:
-                msg = "%s was not found in remote '%s'" % (str(conan_reference), ref_remote.name)
+                msg = "%s was not found in remote '%s'" % (str(conan_reference), remote.name)
                 self._recorder.recipe_install_error(conan_reference, INSTALL_ERROR_MISSING,
-                                                    msg, ref_remote.url)
+                                                    msg, remote.url)
                 raise NotFoundException(msg)
             except RequestException as exc:
                 self._recorder.recipe_install_error(conan_reference, INSTALL_ERROR_NETWORK,
-                                                    str(exc), ref_remote.url)
+                                                    str(exc), remote.url)
                 raise exc
 
         output.info("Not found in local cache, looking in remotes...")
@@ -121,10 +122,10 @@ class ConanProxy(object):
         for remote in remotes:
             logger.debug("Trying with remote %s" % remote.name)
             try:
-                _retrieve_from_remote(remote)
-                return remote
+                new_ref = _retrieve_from_remote(remote)
+                return remote, new_ref
             # If not found continue with the next, else raise
-            except NotFoundException as exc:
+            except NotFoundException:
                 pass
         else:
             msg = "Unable to find '%s' in remotes" % str(conan_reference)
