+++ conans/client/rest/rest_client_v2.py
@@ -1,0 +1,7 @@
+from conans.client.rest.rest_client_common import RestCommonMethods
+
+
+class RestV2Methods(RestCommonMethods):
+
+    def __init__(self, *args, **kwargs):
+        raise NotImplementedError()
