+++ conans/test/remote/auth_bearer_test.py
@@ -1,3 +1,4 @@
+import os
 import unittest
 from conans.test.utils.tools import TestServer, TestClient
 from bottle import request
@@ -58,12 +59,21 @@ class AuthorizeBearerTest(unittest.TestCase):
         errors = client.run("upload Hello/0.1@lasote/stable")
         self.assertFalse(errors)
 
-        expected_calls = [('get_conan_manifest_url', None),
-                          ('check_credentials', None),
-                          ('authenticate', 'Basic'),
-                          ('get_conanfile_snapshot', 'Bearer'),
-                          ('get_conanfile_upload_urls', 'Bearer'),
-                          ('put', None)]
+        if not os.getenv("CONAN_TESTING_SERVER_V2_ENABLED"):
+            expected_calls = [('ping', None),
+                              ('get_conan_manifest_url', None),
+                              ('check_credentials', None),
+                              ('authenticate', 'Basic'),
+                              ('get_conanfile_snapshot', 'Bearer'),
+                              ('get_conanfile_upload_urls', 'Bearer'),
+                              ('put', None)]
+        else:
+            expected_calls = [('ping', None),
+                              ('get_recipe_file', None),
+                              ('check_credentials', None),
+                              ('authenticate', 'Basic'),
+                              ('get_conanfile_snapshot', 'Bearer'),
+                              ('upload_recipe_file', 'Bearer')]
 
         self.assertEqual(len(expected_calls), len(auth.auths))
         for i, (method, auth_type) in enumerate(expected_calls):
@@ -72,6 +82,7 @@ class AuthorizeBearerTest(unittest.TestCase):
             if auth_type:
                 self.assertIn(auth_type, real_call[1])
 
+    @unittest.skipUnless(os.getenv("CONAN_TESTING_SERVER_V2_ENABLED") is None, "ApiV1 test")
     def no_signature_test(self):
         auth = AuthorizationHeaderSpy()
         retur = ReturnHandlerPlugin()
@@ -84,7 +95,8 @@ class AuthorizeBearerTest(unittest.TestCase):
         errors = client.run("upload Hello/0.1@lasote/stable", ignore_error=True)
         self.assertTrue(errors)
 
-        expected_calls = [('get_conan_manifest_url', None),
+        expected_calls = [('ping', None),
+                          ('get_conan_manifest_url', None),
                           ('check_credentials', None),
                           ('authenticate', 'Basic'),
                           ('get_conanfile_snapshot', 'Bearer'),
